Create Assignment 1
Part a)

Create a class called MyDegree. It should have two variables of type double. One called fahrenheit and the other celsius. It should have one constructor that takes in two variables, the first of type double the second of type String. Depending on the value of the String parameter (which should be either fahrenheit or celsius) set the corresponding variable to the value passed as a parameter. Compute the value for the other variable. (Hint: Fahrenheit to Celsius: f = (9/5)*c +32, Celsius to Fahrenheit: c = (5/9)*(f-32). Create 2 get methods, one for each variable. Create 2 set methods, one for each variable (don’t forget to change the other variable from within each set method).

Part b)

Create a class called MyDate. It should have 3 variables of type int: named date, month, and year. It should have 2 variables of type MyDegree: named high and low. Create a constructor that takes in 3 parameters of type int. The constructor should call method isValid and pass it all three parameters. If the 3 parameter were all valid values, they should be set to the date month and year variables. If the isValid method returns false, some default date should be used instead (you choose what the default date value should be). If the default date was used, a JOptionPane message dialog box should be used to alert the user that the entered date was invalid and which date was used instead. Create a method isValid that will verify if the values passed are a valid date. Create set methods for the high and low variables. These 2 set methods should each take in 2 parameters, those 2 parameters should be passed to the constructor of class MyDegree and the resulting object created should be set to the variable (high or low). Create get methods for all 5 variables. The get methods for high and low should not return the MyDegree object, but instead should return the degree value of that object as an int. (Make sure to convert it to an int)

Part c)

Create a class called main which has a main method within it. The main class should ask the user through JOptionPane objects for a date, month, and year. Use 3 JOptionPane objects. Create a MyDate object. Using the get methods of MyDate, verify with the user (using a Confirm Dialog box (a method within JOptionPane) that the MyDate object has the correct date. Remeber, you must get the values from the MyDate object and not use the variables that were passed to the MyDate constructor, since, if the date was not valid, a default date was used instead. If the user clicks the NO button, the user should be asked for another date using JOptionPanes. HINT: use a while loop.  If the user clicks the YES button, the MyDate object should be stored in an array of type Mydate. HINT: create the array outside (ie. before) the while loop, otherwise it will become "out of scope" (ie. cease to exist) after the while loop. Create 4 more MyDate objects for the next 4 days in the calendar. HINT: use the isValid method from class MyDate to determine if these next dates are valid (for example, if the 1st date is June 30th, the 2nd date should be July 1st, not June 31st). Store these other MyDate objects in the array. Then, from within this main method, use the set methods of class MyDate to set the high and low variables. Make up values. Do not try to look up the high and low temperatures of those days. Lastly, print out to the console the high and low temperatures for all 5 days in Fahrenheit and then print out the high and low temperatures in Celsius.

an example output should look like this:

Fahrenheit:

6/30/14 high 84 low 69

7/1/14 high 82 low 61

...

Celsius:

6/30/14 high .... etc.